+++ python/binding/bind_occupancy_trees.cpp
@@ -48,6 +48,25 @@ BindOccupancyTrees(const py::module& m) {
         py::arg("padding") = 0);
     BindOccupancyOctree<OccupancyOctree, OccupancyOctreeNode>(m, "OccupancyOctree", "OccupancyOctreeNode");
 
-    BindOccupancyQuadtree<SurfaceMappingQuadtree, SurfaceMappingQuadtreeNode>(m, "SurfaceMappingQuadtree", "SurfaceMappingQuadtreeNode");
-    BindOccupancyOctree<SurfaceMappingOctree, SurfaceMappingOctreeNode>(m, "SurfaceMappingOctree", "SurfaceMappingOctreeNode");
+    auto [surface_mapping_quadtree, surface_mapping_quadtree_node] =
+        BindOccupancyQuadtree<SurfaceMappingQuadtree, SurfaceMappingQuadtreeNode>(m, "SurfaceMappingQuadtree", "SurfaceMappingQuadtreeNode");
+    py::class_<SurfaceMappingQuadtreeNode::SurfaceData, std::shared_ptr<SurfaceMappingQuadtreeNode::SurfaceData>>(surface_mapping_quadtree_node, "SurfaceData")
+        .def_readwrite("position", &SurfaceMappingQuadtreeNode::SurfaceData::position)
+        .def_readwrite("normal", &SurfaceMappingQuadtreeNode::SurfaceData::normal)
+        .def_readwrite("var_position", &SurfaceMappingQuadtreeNode::SurfaceData::var_position)
+        .def_readwrite("var_normal", &SurfaceMappingQuadtreeNode::SurfaceData::var_normal)
+        .def("__eq__", &SurfaceMappingQuadtreeNode::SurfaceData::operator==)
+        .def("__ne__", &SurfaceMappingQuadtreeNode::SurfaceData::operator!=);
+    surface_mapping_quadtree_node.def_property_readonly("surface_data", &SurfaceMappingQuadtreeNode::GetSurfaceData);
+
+    auto [surface_mapping_octree, surface_mapping_octree_node] =
+        BindOccupancyOctree<SurfaceMappingOctree, SurfaceMappingOctreeNode>(m, "SurfaceMappingOctree", "SurfaceMappingOctreeNode");
+    py::class_<SurfaceMappingOctreeNode::SurfaceData, std::shared_ptr<SurfaceMappingOctreeNode::SurfaceData>>(surface_mapping_octree_node, "SurfaceData")
+        .def_readwrite("position", &SurfaceMappingOctreeNode::SurfaceData::position)
+        .def_readwrite("normal", &SurfaceMappingOctreeNode::SurfaceData::normal)
+        .def_readwrite("var_position", &SurfaceMappingOctreeNode::SurfaceData::var_position)
+        .def_readwrite("var_normal", &SurfaceMappingOctreeNode::SurfaceData::var_normal)
+        .def("__eq__", &SurfaceMappingOctreeNode::SurfaceData::operator==)
+        .def("__ne__", &SurfaceMappingOctreeNode::SurfaceData::operator!=);
+    surface_mapping_octree_node.def_property_readonly("surface_data", &SurfaceMappingOctreeNode::GetSurfaceData);
 }